pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "idx_trading_system"
version = "1.0.0"
description = "Research-grade neural network trading system for Indonesian equities"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Project1 Contributors"}
]
keywords = ["trading", "machine-learning", "neural-networks", "indonesian-stocks", "quantitative-finance"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Financial and Insurance Industry",
    "Topic :: Office/Business :: Financial :: Investment",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "numpy>=1.21.0",
    "pandas>=1.3.0",
    "yfinance>=0.2.0",
    "scikit-learn>=1.0.0",
    "xgboost>=1.5.0",
    "lightgbm>=3.3.0",
    "torch>=1.10.0",
    "pyyaml>=6.0",
    "matplotlib>=3.4.0",
    "seaborn>=0.11.0",
    "joblib>=1.1.0",
    "ta>=0.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=3.0",
    "black>=22.0",
    "flake8>=4.0",
    "mypy>=0.950",
    "jupyter>=1.0",
    "notebook>=6.4",
]

[project.urls]
Homepage = "https://github.com/nicolascheng78/Project1"
Repository = "https://github.com/nicolascheng78/Project1"
Documentation = "https://github.com/nicolascheng78/Project1/blob/main/README.md"

[project.scripts]
idx-backtest = "idx_trading_system.scripts.run_backtest:main"

[tool.setuptools]
packages = ["idx_trading_system"]

[tool.setuptools.package-data]
idx_trading_system = ["config.yaml"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'

[tool.pytest.ini_options]
testpaths = ["idx_trading_system/tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
